Rq 도입하여 쿠키 관련 구현을 대신 처리하게 시킴

+++ src/main/java/com/sbs/tutorial1/boundedContext/member/dto/Member.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @AllArgsConstructor
 @Data
 public class Member {
-  private static int lastId;
+  private static long lastId;
   private final long id;
   private final String username;
   private String password;
